exporter/requirements.txt
aiocontextvars==0.2.2
APScheduler==3.9.1
azt==0.0.7
backports.zoneinfo==0.2.1
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.0.4
contextvars==2.4
cx-Oracle==8.3.0
Cython==0.29.28
dataclasses==0.8
idna==3.3
immutables==0.17
importlib-metadata==4.8.3
importlib-resources==5.4.0
influxdb==5.3.1
loguru==0.6.0
msgpack==1.0.3
pycrypto==2.6.1
pydantic==1.9.0
pyfiglet==0.8.post1
pymssql==2.2.4
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
six==1.16.0
termcolor==1.1.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
zipp==3.6.0
